refactor: rename files to match naming conventions (#107)
Most of these just straight renames.

- "move-to-temp.js" exported two things, so I split that into
  "moveToTemp.js" and "removeFromTemp.js" and simplified them in order
  to reduce duplication (ie. all the stuff that looks up the current
  working directory, joins paths etc). Instead of taking a directory
  parameter and also caching the `process.cwd()` at module evaluation
  time (which itself can cause problems and make testing hard), they
  always just operate in the current directory. At the end, the only
  thing duplicated between the two files is the `TEMP_STRING` constant,
  which is probably ok.

Closes: https://github.com/liferay/liferay-npm-tools/issues/107
