+++ packages/liferay-npm-scripts/bin/liferay-npm-scripts.js
@@ -6,7 +6,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-const {SpawnError} = require('../src/utils/spawn-sync');
+const {SpawnError} = require('../src/utils/spawnSync');
 const log = require('../src/utils/log');
 
 try {
