+++ packages/liferay-npm-scripts/src/scripts/theme.js
@@ -7,7 +7,7 @@
 const fs = require('fs');
 const path = require('path');
 
-const spawnSync = require('../utils/spawn-sync');
+const spawnSync = require('../utils/spawnSync');
 
 const MODULES_DIR = '<RootDir>';
