+++ packages/liferay-npm-scripts/src/utils/deepMerge.js
