+++ packages/liferay-npm-scripts/src/utils/getUserConfig.js
