+++ packages/liferay-npm-scripts/src/utils/setEnv.js
