+++ packages/liferay-npm-scripts/src/utils/spawnSync.js
